pyproject.toml
[project]
name = "base-fastapi"
version = "0.1.0"
description = "Base repo for simple fastapi app"
authors = [
    {name = "dmitrii_mishchenko",email = "dmittry.m@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "redis (>=6.1.0,<7.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
black = "^25.1.0"
isort = "^6.0.1"
ruff = "^0.11.10"
pre-commit = "^4.2.0"
pytest = "^8.3.5"
pytest-asyncio = "^1.0.0"
pytest-redis = "^3.1.3"
freezegun = "^1.5.2"

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py313"
fix = false
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

[tool.ruff.lint]
select = ["E", "F", "W", "C", "B", "SIM", "I", "D"]
exclude = ["tests/*", "app/migrations/*"]
ignore = ["D104", "D100", 'D205', "D211", "D213", "I001", "D203"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.mypy]
files = "app"
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
explicit_package_bases = true

[tool.isort]
profile = "black"
line_length = 88
known_first_party = [ "utils", "api", "core", "tests", "services", 'infrastructure']
known_third_party = ['fastapi', 'sqlalchemy', 'pydantic', 'pydantic-settings', "pyjwt", "redis", "starlette"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip = ["migrations"]

[tool.black]
line-length = 88
